Presentation & Functionality
Functionality of Application from a users perspective. How using InterStock can be useful to a seasoned Investor. By searching for Index's in International Markets and Exchanges within the same economic region that work in conjunction with one another. An Investor conveniently search each potential trade index within each others Exchange or Indices of origin. To be tracked, compared and liquidated for the best short or long term Investment Outcome.